feat(issue-134): edit post (#153)

+++ src/pages/api/app/[appAbbreviation]/album/post/[postId].ts
@@ -35,23 +35,24 @@ export default async function handler(
       return res.status(401).json({ message: 'SEM_NOT_AUTHORIZED_USER' })
     }
     const currentUserAppAccessLevel = currentUser.accessLevel[appAbbreviation]
-    // if the user access-level is not super admin or admin, return error
-    if (
-      currentUserAppAccessLevel !== AccessLevels.SUPER_ADMIN &&
-      currentUserAppAccessLevel !== AccessLevels.ADMIN
-    ) {
-      return res.status(401).json({ message: 'SEM_NOT_AUTHORIZED_USER' })
-    }
 
     const postDoc = createDocument(CollectionsName.ALBUM_POST, categorySchema)
 
     switch (req.method) {
       case 'PUT': {
-        const { audience, categoriesId, title, description } = req.body ?? {}
+        // if the user access-level is not super admin or admin, return error
+        if (
+          currentUserAppAccessLevel !== AccessLevels.SUPER_ADMIN &&
+          currentUserAppAccessLevel !== AccessLevels.ADMIN
+        ) {
+          return res.status(401).json({ message: 'SEM_NOT_AUTHORIZED_USER' })
+        }
+
+        const { audience, categories, title, description } = req.body ?? {}
 
         const post: PostRaw | null = await postDoc.findOneAndUpdate(
           { _id: postId },
-          { audience, categoriesId, title, description },
+          { audience, categories, title, description },
           { new: true, runValidators: true }
         )
 
@@ -63,6 +64,11 @@ export default async function handler(
       }
 
       case 'DELETE': {
+        // if the user access-level is not super admin, return error
+        if (currentUserAppAccessLevel !== AccessLevels.SUPER_ADMIN) {
+          return res.status(401).json({ message: 'SEM_NOT_AUTHORIZED_USER' })
+        }
+
         const deletedPost: PostRaw | null = await postDoc.findOneAndDelete({
           _id: postId
         })
